Completly revamped project structure, now using dynamic libraries, new shared module and multiple code improvements

+++ luly/core/src/input/input_manager.cpp
@@ -1,22 +1,22 @@
 ﻿#include "lypch.h"
-#include "input_manager.h"
+
+#include <input/input_manager.h>
 
 #include "application/application.h"
-#include "window/window.h"
 
 namespace luly::input
 {
     bool input_manager::is_key_pressed(key key_code)
     {
         GLFWwindow* window = core::application::get().get_window()->get_native_handle();
-        const int state = glfwGetKey(window, key_code);
+        const int state = glfwGetKey(window, static_cast<int>(key_code));
         return state == GLFW_PRESS;
     }
 
     bool input_manager::is_mouse_button_pressed(mouse_button button_code)
     {
         GLFWwindow* window = core::application::get().get_window()->get_native_handle();
-        const int state = glfwGetMouseButton(window, button_code);
+        const int state = glfwGetMouseButton(window, static_cast<int>(button_code));
         return state == GLFW_PRESS;
     }
